src/super_resolution.py
"""
Super-resolution module for image enhancement using deep learning models.

This module provides functionality to upscale low-resolution images using
state-of-the-art super-resolution models including ESPCN, Real-ESRGAN, and SwinIR.
"""

import logging
from pathlib import Path
from typing import Tuple, Union, Optional, List
import warnings

import torch
import torch.nn as nn
import torch.nn.functional as F
import torchvision.transforms as transforms
from PIL import Image
import numpy as np
import cv2

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class ESPCN(nn.Module):
    """
    Efficient Sub-Pixel Convolutional Neural Network (ESPCN) implementation.
    
    This is a custom implementation of ESPCN for super-resolution.
    """
    
    def __init__(self, scale_factor: int = 3):
        """
        Initialize ESPCN model.
        
        Args:
            scale_factor: Upscaling factor (2, 3, 4, or 8)
        """
        super(ESPCN, self).__init__()
        self.scale_factor = scale_factor
        
        # Feature extraction layers
        self.conv1 = nn.Conv2d(1, 64, kernel_size=5, padding=2)
        self.conv2 = nn.Conv2d(64, 32, kernel_size=3, padding=1)
        
        # Sub-pixel convolution layer
        self.conv3 = nn.Conv2d(32, scale_factor * scale_factor, kernel_size=3, padding=1)
        
        # Activation functions
        self.relu = nn.ReLU(inplace=True)
        
    def forward(self, x):
        """Forward pass through the network."""
        # Feature extraction
        x = self.relu(self.conv1(x))
        x = self.relu(self.conv2(x))
        
        # Sub-pixel convolution
        x = self.conv3(x)
        
        # Pixel shuffle to upscale
        x = F.pixel_shuffle(x, self.scale_factor)
        
        return x


class SuperResolutionModel:
    """
    A wrapper class for super-resolution models with enhanced functionality.
    
    Supports multiple model architectures and provides a unified interface
    for image upscaling with various preprocessing and postprocessing options.
    """
    
    def __init__(
        self, 
        model_type: str = "espcn", 
        scale_factor: int = 3,
        device: Optional[str] = None
    ):
        """
        Initialize the super-resolution model.
        
        Args:
            model_type: Type of model to use ('espcn', 'real_esrgan', 'swinir')
            scale_factor: Upscaling factor (2, 3, 4, or 8)
            device: Device to run inference on ('cpu', 'cuda', or None for auto)
        """
        self.model_type = model_type.lower()
        self.scale_factor = scale_factor
        self.device = device or ("cuda" if torch.cuda.is_available() else "cpu")
        
        logger.info(f"Initializing {self.model_type} model with scale factor {scale_factor}")
        logger.info(f"Using device: {self.device}")
        
        self.model = self._load_model()
        self.model.eval()
        
    def _load_model(self) -> nn.Module:
        """Load the specified super-resolution model."""
        if self.model_type == "espcn":
            return self._load_espcn_model()
        elif self.model_type == "real_esrgan":
            return self._load_real_esrgan_model()
        elif self.model_type == "swinir":
            return self._load_swinir_model()
        else:
            raise ValueError(f"Unsupported model type: {self.model_type}")
    
    def _load_espcn_model(self) -> nn.Module:
        """Load custom ESPCN model."""
        try:
            if self.scale_factor not in [2, 3, 4, 8]:
                raise ValueError(f"ESPCN only supports scale factors 2, 3, 4, 8. Got {self.scale_factor}")
            
            model = ESPCN(scale_factor=self.scale_factor)
            
            # Initialize weights (Xavier initialization)
            for m in model.modules():
                if isinstance(m, nn.Conv2d):
                    nn.init.xavier_uniform_(m.weight)
                    if m.bias is not None:
                        nn.init.constant_(m.bias, 0)
            
            logger.info(f"Initialized custom ESPCN model with scale factor {self.scale_factor}")
            return model.to(self.device)
        except Exception as e:
            logger.error(f"Failed to load ESPCN model: {e}")
            raise
    
    def _load_real_esrgan_model(self) -> nn.Module:
        """Load Real-ESRGAN model (placeholder for future implementation)."""
        logger.warning("Real-ESRGAN model not yet implemented, falling back to ESPCN")
        return self._load_espcn_model()
    
    def _load_swinir_model(self) -> nn.Module:
        """Load SwinIR model (placeholder for future implementation)."""
        logger.warning("SwinIR model not yet implemented, falling back to ESPCN")
        return self._load_espcn_model()
    
    def preprocess_image(self, image_path: Union[str, Path]) -> Tuple[torch.Tensor, Image.Image, Image.Image, Image.Image]:
        """
        Preprocess image for super-resolution.
        
        Args:
            image_path: Path to the input image
            
        Returns:
            Tuple containing:
            - input_tensor: Preprocessed tensor for model input
            - cb: Cb channel of original image
            - cr: Cr channel of original image  
            - original_image: Original RGB image
        """
        try:
            image_path = Path(image_path)
            if not image_path.exists():
                raise FileNotFoundError(f"Image not found: {image_path}")
            
            # Load and convert image
            image = Image.open(image_path).convert('RGB')
            original_size = image.size
            
            # Ensure dimensions are divisible by scale factor
            new_width = (image.width // self.scale_factor) * self.scale_factor
            new_height = (image.height // self.scale_factor) * self.scale_factor
            
            if new_width != image.width or new_height != image.height:
                logger.info(f"Resizing image from {original_size} to ({new_width}, {new_height})")
                image = image.resize((new_width, new_height), Image.LANCZOS)
            
            # Convert to YCbCr color space
            ycbcr = image.convert('YCbCr')
            y, cb, cr = ycbcr.split()
            
            # Convert Y channel to tensor
            transform = transforms.ToTensor()
            input_tensor = transform(y).unsqueeze(0).to(self.device)
            
            return input_tensor, cb, cr, image
            
        except Exception as e:
            logger.error(f"Error preprocessing image {image_path}: {e}")
            raise
    
    def upscale_image(self, image_path: Union[str, Path]) -> Tuple[Image.Image, Image.Image]:
        """
        Upscale a low-resolution image using the loaded model.
        
        Args:
            image_path: Path to the input image
            
        Returns:
            Tuple containing (original_image, upscaled_image)
        """
        try:
            input_tensor, cb, cr, original_image = self.preprocess_image(image_path)
            
            with torch.no_grad():
                output = self.model(input_tensor)
            
            # Postprocess the output
            output = output.squeeze().clamp(0, 1).cpu()
            out_y = transforms.ToPILImage()(output)
            
            # Upscale Cb and Cr channels to match the upscaled Y channel
            cb_up = cb.resize(out_y.size, Image.BICUBIC)
            cr_up = cr.resize(out_y.size, Image.BICUBIC)
            
            # Merge channels back to RGB
            final_image = Image.merge("YCbCr", [out_y, cb_up, cr_up]).convert("RGB")
            
            logger.info(f"Successfully upscaled image from {original_image.size} to {final_image.size}")
            return original_image, final_image
            
        except Exception as e:
            logger.error(f"Error upscaling image {image_path}: {e}")
            raise
    
    def batch_upscale(self, image_paths: List[Union[str, Path]]) -> List[Tuple[Image.Image, Image.Image]]:
        """
        Upscale multiple images in batch.
        
        Args:
            image_paths: List of paths to input images
            
        Returns:
            List of tuples containing (original_image, upscaled_image) for each input
        """
        results = []
        for i, image_path in enumerate(image_paths):
            try:
                logger.info(f"Processing image {i+1}/{len(image_paths)}: {image_path}")
                result = self.upscale_image(image_path)
                results.append(result)
            except Exception as e:
                logger.error(f"Failed to process {image_path}: {e}")
                results.append(None)
        
        return results


class ImageQualityMetrics:
    """Calculate image quality metrics for super-resolution evaluation."""
    
    @staticmethod
    def calculate_psnr(img1: np.ndarray, img2: np.ndarray) -> float:
        """Calculate Peak Signal-to-Noise Ratio (PSNR)."""
        mse = np.mean((img1 - img2) ** 2)
        if mse == 0:
            return float('inf')
        return 20 * np.log10(255.0 / np.sqrt(mse))
    
    @staticmethod
    def calculate_ssim(img1: np.ndarray, img2: np.ndarray) -> float:
        """Calculate Structural Similarity Index (SSIM)."""
        # Simple SSIM implementation - in production, use scikit-image
        mu1 = np.mean(img1)
        mu2 = np.mean(img2)
        sigma1 = np.var(img1)
        sigma2 = np.var(img2)
        sigma12 = np.mean((img1 - mu1) * (img2 - mu2))
        
        c1 = 0.01 ** 2
        c2 = 0.03 ** 2
        
        ssim = ((2 * mu1 * mu2 + c1) * (2 * sigma12 + c2)) / \
               ((mu1 ** 2 + mu2 ** 2 + c1) * (sigma1 + sigma2 + c2))
        
        # Clamp SSIM to valid range [-1, 1]
        return max(-1.0, min(1.0, ssim))


def create_sample_dataset(output_dir: Union[str, Path], num_samples: int = 10) -> List[Path]:
    """
    Create a synthetic dataset of low-resolution images for testing.
    
    Args:
        output_dir: Directory to save sample images
        num_samples: Number of sample images to create
        
    Returns:
        List of paths to created sample images
    """
    output_dir = Path(output_dir)
    output_dir.mkdir(parents=True, exist_ok=True)
    
    created_files = []
    
    for i in range(num_samples):
        # Create a synthetic image with some patterns
        width, height = 120, 120  # Low resolution
        
        # Create a simple pattern
        img_array = np.random.randint(0, 256, (height, width, 3), dtype=np.uint8)
        
        # Add some geometric patterns
        center_x, center_y = width // 2, height // 2
        y, x = np.ogrid[:height, :width]
        
        # Create concentric circles
        mask = ((x - center_x) ** 2 + (y - center_y) ** 2) < (min(width, height) // 3) ** 2
        img_array[mask] = [255, 100, 100]  # Red circle
        
        # Add some noise
        noise = np.random.randint(-30, 31, (height, width, 3))
        img_array = np.clip(img_array.astype(np.int16) + noise, 0, 255).astype(np.uint8)
        
        # Save image
        img = Image.fromarray(img_array)
        file_path = output_dir / f"sample_{i:03d}.jpg"
        img.save(file_path, "JPEG", quality=85)
        created_files.append(file_path)
    
    logger.info(f"Created {num_samples} sample images in {output_dir}")
    return created_files


if __name__ == "__main__":
    # Example usage
    model = SuperResolutionModel(model_type="espcn", scale_factor=3)
    
    # Create sample dataset
    sample_images = create_sample_dataset("data/samples", num_samples=5)
    
    # Process first sample
    if sample_images:
        original, upscaled = model.upscale_image(sample_images[0])
        print(f"Original size: {original.size}")
        print(f"Upscaled size: {upscaled.size}")
