requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0

# Web interface
streamlit>=1.28.0

# Configuration and utilities
PyYAML>=6.0
pathlib2>=2.3.0

# Image processing and visualization
matplotlib>=3.5.0
scikit-image>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional: Advanced super-resolution models
# realesrgan>=0.3.0  # Uncomment for Real-ESRGAN support
# basicsr>=1.4.0     # Uncomment for BasicSR support

# Optional: Additional image formats
# imageio>=2.19.0
# tifffile>=2022.5.4
